Refactor: move session to src, decouple IO dependencies
